Extracted log functions from the views
